+++ src/MainWindow.cpp
@@ -9,8 +9,8 @@
 #include "Toolbar.h"
 #include "GuiConfig.h"
 #include "DDEProxyModeManager.h"
-#include "ProxyView.h"
-#include "ConfigItem.h"
+#include "ProfileView.h"
+#include "ProfileItem.h"
 #include "QRCodeCapturer.h"
 #include "SSValidator.h"
 #include "ShareDialog.h"
