--- src/ProxyThread.cpp
@@ -1,11 +1,0 @@
-#include "ProxyThread.h"
-
-ProxyThread::ProxyThread(QSS::Controller *controller) {
-    this->controller = controller;
-}
-
-void ProxyThread::run() {
-    qDebug() << "start proxy";
-    auto flag = controller->start();
-    qDebug() << "result" << flag;
-}
